+++ src/org/xjge/test/SceneModel.java
@@ -95,7 +95,7 @@ public class SceneModel extends Scene {
     }
 
     @Override
-    public void render(Map<String, Shader> glPrograms, int viewportID, Camera camera, int depthTexHandle) {
+    public void render(int viewportID, Camera camera, int depthTexHandle) {
         for(var entity : queryEntities(Prism.class)) {
             entity.getComponent(Prism.class).render(camera);
         }