+++ src/org/xjge/test/SceneRemove.java
@@ -31,7 +31,7 @@ public class SceneRemove extends Scene {
     }
 
     @Override
-    public void render(Map<String, Shader> glPrograms, int viewportID, Camera camera, int depthTexHandle) {
+    public void render(int viewportID, Camera camera, int depthTexHandle) {
     }
 
     @Override